refactor

+++ database/migrations/2025_12_11_214502_create_tasks_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('tasks', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedSmallInteger('position')->default(0);
+            $table->unsignedTinyInteger('position')->default(0);
             $table->boolean('completed')->default(false);
             $table->timestamps();
         });
